+++ games/urls.py
@@ -30,6 +30,7 @@ urlpatterns = [
     ),
     path("add-purchase/", views.add_purchase, name="add_purchase"),
     path("add-edition/", views.add_edition, name="add_edition"),
+    path("add-device/", views.add_device, name="add_device"),
     path("edit-session/<int:session_id>", views.edit_session, name="edit_session"),
     path("list-sessions/", views.list_sessions, name="list_sessions"),
     path(